// front/src/app/models/result.ts
export class Result {
  id: number;
  result: number;
  periodStr1: number;
  amountFromSaleStr2: number;
  amountNoOperInStr3: number;
  placeToWorkStr4: boolean;
  existBenefitsStr5: boolean;
  widowSingleFosterGuardianStr6: boolean;
  numberChildUnderEighteenStr7: number;
  numberDisableChildStr8: number;
  numberDependentsStr9: number;
  amountInsurancePemiumStr10: number;
  amountGetFirstPaidEducationStr11: number;
  amountForConstructHouseStr12: number;
  amountEntreprenActivitiesStr13: number;

}